Make locality scheduler first-touch aware
The locality scheduler now consideres pages that have not yet been
first-touched. If the number of bytes in a task dependency that
have not yet been first-touched exceeds the number of bytes backed
by one single node, multiplied by a tuning parameter, the node
to offload to is selected in a cyclical fashion instead of by the
default locality policy. The tuning parameter argo.locality_tuning
can be found in the nanos6.toml file.
This commit also fixes a bug causing unnecessary self-downgrades
on task offload.