Introduce argodsm.weak_release optimisation
This configuration parameter enables nanos6 with ArgoDSM support
to only perform self-downgrade on weak tasks. This means that
strong tasks must be wrapped in weak tasks with the same or greater
dependencies, and that the strong tasks must be waited for, as
they do not perform self-downgrade on their own.